Rename validation map file
